move Combine helpers to a new file
